Draft Dropdown component with headless-ui layers
