+++ pkg/ui/src/Dropdown/Dropdown.story.svelte
@@ -2,7 +2,7 @@
 	import type { Hst } from '@histoire/plugin-svelte';
 	import { Settings2 } from 'lucide-svelte';
 
-	import { DropdownTrigger, DropdownItems, Dropdown } from './index';
+	import { Dropdown } from './index';
 
 	export let Hst: Hst;
 
@@ -17,25 +17,7 @@
 </script>
 
 <Hst.Story title="Dropdown" layout={{ type: 'grid' }}>
-	<Hst.Variant title="Dropdown Trigger">
-		<DropdownTrigger label="Options" />
-	</Hst.Variant>
-	<Hst.Variant title="Dropdown Trigger With Icon">
-		<DropdownTrigger label="Options">
-			<div slot="icon"><Settings2 /></div>
-		</DropdownTrigger>
-	</Hst.Variant>
-	<Hst.Variant title="Dropdown Items">
-		<div class="relative h-80">
-			<DropdownItems {items} />
-		</div>
-	</Hst.Variant>
-	<Hst.Variant title="Dropdown Menu">
-		<div>
-			<Dropdown>
-				<DropdownTrigger slot="trigger" label="Options" />
-				<DropdownItems slot="items" {items} />
-			</Dropdown>
-		</div>
-	</Hst.Variant>
+	<Dropdown buttonLabel="Options" {items}>
+		<div slot="buttonIcon"><Settings2 /></div>
+	</Dropdown>
 </Hst.Story>
