+++ pkg/ui/src/Dropdown/index.ts
@@ -1,5 +1,3 @@
-export { default as DropdownTrigger } from './DropdownTrigger.svelte';
-export { default as DropdownItems } from './DropdownItems.svelte';
 export { default as Dropdown } from './Dropdown.svelte';
 
 export * from './enums';
